Service/requirements.txt
Flask==2.3.2
locustio==0.999
numpy==1.23.2
Pillow==9.5.0
requests==2.28.1
torch==2.0.1+cu117
torchvision==0.15.2+cu117
